feat: :sparkles: Fixed some linting errors.
