+++ package.json
@@ -30,6 +30,7 @@
     "@vitejs/plugin-react": "^3.1.0",
     "eslint": "^8.35.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-plugin-react": "^7.32.2",
     "prettier": "2.8.4",
     "tailwindcss": "^3.2.7",
     "typescript": "^4.9.3",
