+++ src/App.tsx
@@ -1,9 +1,10 @@
-import './App.css'
-import { HomePage } from './components/pages/HomePage'
+import './App.css';
+import { HomePage } from './components/pages/HomePage';
+
 function App() {
   return (
     <HomePage />
-    )
+  );
 }
 
-export default App
+export default App;
