+++ src/components/UI/molecules/ListComponent.tsx
@@ -1,15 +1,10 @@
-import React from 'react'
-import { IListComponent } from '../../../types/types'
-import { ListCard } from '../atoms/ListCard';
-import {List} from 'antd'
+import React from 'react';
+import { List } from 'antd';
+import { IListComponent } from '../../../types/types';
 
 export const ListComponent = ({
-    children,
-    skeletonLoader,
-    styles
-}: IListComponent) => {
-    return (
+  styles,
+}: IListComponent) => (
         <List className={`${styles} h-full overflow-auto`}>
         </List>
-    );
-}
+);
